Prebid 10: rename fanAdapter to fanBidAdapter (#13371)
