+++ modules/fanBidAdapter.js
