+++ modules/fanBidAdapter.md
